update ui profile and admin panel

+++ SportsHub/frontend/src/pages/Logout.jsx
@@ -1,16 +1,26 @@
 import React from 'react'
 import { useEffect } from 'react'
 import { useNavigate } from 'react-router'
+import { showCustomToast } from '../helper/CustomToast';
+import axios from 'axios';
 function Logout({setIsLoggedIn}) {
   const navigate=useNavigate();
   useEffect(()=>{
-    setIsLoggedIn(false);
-    localStorage.clear();
-      navigate("/login");
+    const log_out= async ()=>{
+      const res=await axios.get("http://localhost:3000/users/logout",{withCredentials:true})
+      if(res.status==200){
+        setIsLoggedIn(false);
+        localStorage.clear();
+        navigate("/login");
+      }
+      else{
+        showCustomToast("error","check you network connection")
+      }
+    }
+    log_out()
   })
   return (
     <div>
-      
     </div>
   )
 }
